OOjs Router method to provide access to replaceState
This simple change enables frontend developers with a way for
invoking history.replaceState safely. This will prove extremely
useful in MultimediaViewer and MobileFrontend.

A new function navigateTo is added for backwards compatibility
with MobileFrontend. The navigate method is now deprecated.

Given history API can assume to be supported in all grade A
browsers the checks are removed and history support is assumed.

Tests are added for all cases - new and old.

Bug: T189173
Change-Id: I1ad0579b1336e6e83d00fa26067b5072de718a28