Update patch set 4
Patch Set 4:

Afaik history.pushState is supported in a superset of browsers we support Grade A for (which is feature-test based), which means JS does not execute in browsers where this capability is absent.

It might be easier for developers to work with the History API directly, which would also make the code more portable and benefits from e.g. discoverable MDN docs based on searching for the interface seen in code, so we don't have to maintain a discoverability path from OO.Router > (something T.B.D.) > MDN docs etc.

Fresh on my mind from re-reading this blogpost over the holidays, where one could substitute "jQuery" for "OO.Router" in some of the later paragraphs:

https://githubengineering.com/removing-jquery-from-github-frontend/

Patch-set: 4
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>