Update patch set 2
Patch Set 2:

Ah, yeah, that's not a supported way of calling the method and could break without warning in a future version.

I'd recommend having the OO.Router instance available earlier and re-using it between the code paths that need it, or to simply use location.hash.slice(1) directly there if you want it to be independent from the router instance. Using a static method wouldn't really make sense because the encapsulation here is meant to offer abstraction. But, if it's unable to use instance state, and can only return one thing, it's basically just a utility method that we can never change, and no longer related to OO.Router.

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>